refactor: changing nextauth version to latest stable and fixing broken auth functions Now the user can sign in with oauth or with a custom user, and the server component will decode the token and correctly render the header
