+++ frontend/package.json
@@ -13,7 +13,7 @@
     "bcrypt": "^5.1.1",
     "jsonwebtoken": "^9.0.2",
     "next": "14.2.8",
-    "next-auth": "5.0.0-beta.25",
+    "next-auth": "^4.24.8",
     "react": "^18",
     "react-dom": "^18",
     "react-hook-form": "^7.53.0"
